Add separate label for "super" sponsors
We want to make a subtle difference when sponsors pitched in above a certain amount, which is the trigger for prioritizing their issues. This is made by just changing the color of the heart in the label.

In addition, this fixes the label background color to better match the GH sponsors color.